+++ src/Middleware/VanillaMiddleware.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Rougin\Slytherin\Middleware\Vanilla;
+
+/**
+ * Vanilla Middleware
+ *
+ * A simple implementation of a middleware.
+ * NOTE: To be removed in v1.0.0
+ *
+ * @package Slytherin
+ * @author  Rougin Royce Gutib <rougingutib@gmail.com>
+ */
+class VanillaMiddleware implements Vanilla\Middleware
+{
+}