+++ tests/Middleware/Stratigility/MiddlewareTest.php
@@ -21,21 +21,24 @@ class MiddlewareTest extends \PHPUnit_Framework_TestCase
             $this->markTestSkipped('Zend Stratigility is not installed.');
         }
 
-        $queue = array();
+        $_SERVER['REQUEST_METHOD'] = 'GET';
+        $_SERVER['REQUEST_URI']    = '/';
+        $_SERVER['SERVER_NAME']    = 'localhost';
+        $_SERVER['SERVER_PORT']    = '8000';
 
-        array_push($queue, 'Rougin\Slytherin\Fixture\Middlewares\FirstMiddleware');
-        array_push($queue, 'Rougin\Slytherin\Fixture\Middlewares\SecondMiddleware');
-        array_push($queue, 'Rougin\Slytherin\Fixture\Middlewares\LastMiddleware');
+        $stack = array();
+
+        array_push($stack, 'Rougin\Slytherin\Fixture\Middlewares\FirstMiddleware');
+        array_push($stack, 'Rougin\Slytherin\Fixture\Middlewares\SecondMiddleware');
+        array_push($stack, 'Rougin\Slytherin\Fixture\Middlewares\LastMiddleware');
 
         $pipeline   = new \Zend\Stratigility\MiddlewarePipe;
         $middleware = new \Rougin\Slytherin\Middleware\Stratigility\Middleware($pipeline);
 
-        $stream   = new \Rougin\Slytherin\Http\Stream;
-        $uri      = new \Rougin\Slytherin\Http\Uri;
-        $request  = new \Rougin\Slytherin\Http\ServerRequest('1.1', array(), $stream, '/', 'GET', $uri);
+        $request  = new \Rougin\Slytherin\Http\ServerRequest($_SERVER);
         $response = new \Rougin\Slytherin\Http\Response;
 
-        $response = $middleware($request, $response, $queue);
+        $response = $middleware($request, $response, $stack);
 
         $this->assertEquals('First! Second! Last!', (string) $response->getBody());
     }