+++ tests/Middleware/Vanilla/MiddlewareTest.php
@@ -17,21 +17,46 @@ class MiddlewareTest extends \PHPUnit_Framework_TestCase
      */
     public function testInvokeMethod()
     {
-        $queue = array();
-
-        array_push($queue, 'Rougin\Slytherin\Fixture\Middlewares\FirstMiddleware');
-        array_push($queue, 'Rougin\Slytherin\Fixture\Middlewares\SecondMiddleware');
-        array_push($queue, 'Rougin\Slytherin\Fixture\Middlewares\LastMiddleware');
+        $_SERVER['REQUEST_METHOD'] = 'GET';
+        $_SERVER['REQUEST_URI']    = '/';
+        $_SERVER['SERVER_NAME']    = 'localhost';
+        $_SERVER['SERVER_PORT']    = '8000';
 
         $middleware = new \Rougin\Slytherin\Middleware\Vanilla\Middleware;
+        $request    = new \Rougin\Slytherin\Http\ServerRequest($_SERVER);
+        $response   = new \Rougin\Slytherin\Http\Response;
 
-        $stream   = new \Rougin\Slytherin\Http\Stream;
-        $uri      = new \Rougin\Slytherin\Http\Uri;
-        $request  = new \Rougin\Slytherin\Http\ServerRequest('1.1', array(), $stream, '/', 'GET', $uri);
-        $response = new \Rougin\Slytherin\Http\Response;
+        $middleware->push(function ($request, $next = null) {
+            return $next($request);
+        });
 
-        $response = $middleware($request, $response, $queue);
+        $middleware->push('Rougin\Slytherin\Fixture\Middlewares\FirstMiddleware');
+        $middleware->push('Rougin\Slytherin\Fixture\Middlewares\SecondMiddleware');
+        $middleware->push('Rougin\Slytherin\Fixture\Middlewares\LastMiddleware');
+
+        $response = $middleware($request, $response, $middleware->getQueue());
 
         $this->assertEquals('First! Second! Last!', (string) $response->getBody());
+    }
+
+    /**
+     * Tests process() method
+     *
+     * @return void
+     */
+    public function testProcessMethod()
+    {
+        $_SERVER['REQUEST_METHOD'] = 'GET';
+        $_SERVER['REQUEST_URI']    = '/';
+        $_SERVER['SERVER_NAME']    = 'localhost';
+        $_SERVER['SERVER_PORT']    = '8000';
+
+        $middleware = new \Rougin\Slytherin\Middleware\Vanilla\Middleware;
+        $request    = new \Rougin\Slytherin\Http\ServerRequest($_SERVER);
+
+        $middleware->push('Rougin\Slytherin\Fixture\Middlewares\InteropMiddleware');
+        $middleware->push('Rougin\Slytherin\Middleware\FinalResponse');
+
+        $this->assertInstanceOf('Psr\Http\Message\ResponseInterface', $middleware->dispatch($request));
     }
 }